Move extra columns binding to Test
