+++ test/Common/ProductExtraColumns.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-namespace Microsoft.Azure.WebJobs.Extensions.Sql.Samples.Common
+namespace Microsoft.Azure.WebJobs.Extensions.Sql.Tests.Common
 {
     public class ProductExtraColumns
     {
